+++ src/ql/arch/cc/pass/gen/vq1asm/detail/options.h
@@ -18,6 +18,7 @@
 #define OPT_SUPPORT_STATIC_CODEWORDS    1   // support (currently: require) static codewords, instead of allocating them on demand
 #define OPT_STATIC_CODEWORDS_ARRAYS     1   // JSON field static_codeword_override is an array with one element per qubit parameter
 #define OPT_VECTOR_MODE                 0   // 1=generate single code word for all output groups together (requires codewords allocated by backend)
+#define OPT_CC_USER_FUNCTIONS           1   // 1=add support for some user functions (experimental)
 
 // NOTE JvS: I added these to the documentation string for the backend pass, so
 //  you can see these values at runtime.